Anime Recommendation System/backend/requirements.txt
langchain==0.3.27
langchain-core==0.3.78
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-groq
langchain-chroma
langchain-google-genai
sentence-transformers
chromadb
faiss-cpu
pypdf
PyPDF2
python-dotenv
streamlit
beautifulsoup4 
fastapi
uvicorn
rank-bm25